optionssearchactivity mostly working

+++ app/src/main/java/ca/sog/SOG/ChainPerStrike.kt
@@ -1,7 +1,7 @@
 package ca.sog.SOG
 
 data class ChainPerStrikePrice(
-        val strikePrice: Int,
+        val strikePrice: Double,
         val callSymbolId: Int,
         val putSymbolId: Int
 )
